Removed _2 _3 weirdness for yacc and lex
Previous versions of this library apparently had problems with bison 2/3.
To solve this, there were two versions for each yy and ll file, one for each version.
This caused an enourmous amount of confusion, and did not seem to be used.
These files have been removed (using the _3 files as the base yy/ll files) and the build process updated.

It should suffice to require bison 3.